+++ src/common/workspace.ts
@@ -1,10 +1,13 @@
 // https://code.visualstudio.com/docs/editor/workspaces
 
 import * as path from 'path';
-import { ConfigurationChangeEvent, ConfigurationScope, Uri, WorkspaceFolder, workspace } from 'vscode'
+import {
+    ConfigurationChangeEvent, ConfigurationScope, Uri,
+    WorkspaceFolder, workspace
+} from 'vscode'
 import { ExtensionTopLevelSection } from '../common'
 
-export function basePathForFilename(uri:Uri): string {
+export function basePathForFilename(uri: Uri): string {
     if (workspace.workspaceFolders) {
         return workspace.getWorkspaceFolder(uri).uri.fsPath;
     } else {
@@ -16,6 +19,6 @@ export function scope(): ConfigurationScope {
 }
 
 // true if config change event effects section 
-export function affectsConfiguration(event:ConfigurationChangeEvent,section:string):boolean {
+export function affectsConfiguration(event: ConfigurationChangeEvent, section: string): boolean {
     return event.affectsConfiguration(`${ExtensionTopLevelSection}.${section}`);
 }