+++ src/providers/activate.ts
@@ -1,16 +1,18 @@
 import { ExtensionContext} from "vscode";
-import {XQLinter} from "../xqlints";
+import {XQLinters} from "../xqlints";
 import * as formatter from "./formatting";
 import * as symbols from './symbols';
 import * as hover from './hover';
+import * as definition from './definition';
 import * as codeAction from './codeaction';
 import * as completion from './completion';
 import * as documentLink from './documentlink';
 
-export function activate(context: ExtensionContext,diagnosticCollectionXQuery:XQLinter) {
+export function activate(context: ExtensionContext,diagnosticCollectionXQuery:XQLinters) {
     symbols.activate(context);
     hover.activate(context);
-    codeAction.activate(context,diagnosticCollectionXQuery);
+    definition.activate(context);
+//  codeAction.activate(context,diagnosticCollectionXQuery);
     completion.activate(context);
     documentLink.activate(context);
     formatter.activate(context);