+++ src/providers/documentlink.ts
@@ -1,9 +1,9 @@
 // XQuery Document link provider
 
-/*---------------------------------------------------------------------------------------------
+/*----------------------------------------------------------
  *  Copyright (c) Quodatum. All rights reserved.
  *  Licensed under the MIT License.
- *--------------------------------------------------------------------------------------------*/
+ *----------------------------------------------------------*/
 
 import {  DocLink } from '@quodatum/xqlint';
 import {
@@ -11,7 +11,7 @@ import {
     DocumentLink, DocumentLinkProvider,Uri
 } from 'vscode';
 import { channel,importRange } from "../common";
-import { diagnosticCollectionXQuery } from "../extension";
+import { xqLinters } from "../extension";
 import { languageIds } from "../constants";
 
 
@@ -24,7 +24,7 @@ class XQueryDocumentLinks implements DocumentLinkProvider {
     // eslint-disable-next-line @typescript-eslint/no-unused-vars
     provideDocumentLinks = async (doc: TextDocument, _token: CancellationToken): Promise<DocumentLink[]> => {
         channel.start("Doclinks" , doc.uri);
-        const linter =   diagnosticCollectionXQuery.xqlint(doc.uri); 
+        const linter =   xqLinters.xqlint(doc.uri); 
         const dlinks = linter.getDocLinks();
         const links: DocumentLink[] = [];
         dlinks.forEach((dl)=>{