+++ src/providers/hover.ts
@@ -2,10 +2,10 @@
 
 import * as vscode from "vscode";
 
-import { Configuration ,inspectAst } from "../common";
+import { Configuration, inspectAst, markdownString } from "../common";
 import { languageIds } from "../constants";
-import { diagnosticCollectionXQuery } from "../extension";
-import { XQLint } from '@quodatum/xqlint';
+import { xqLinters } from "../extension";
+
 
 
 export function activate(context: vscode.ExtensionContext) {
@@ -13,29 +13,36 @@ export function activate(context: vscode.ExtensionContext) {
         { language: languageIds.xquery }, new XQueryHoverProvider()
     ));
 }
+
+
 class XQueryHoverProvider implements vscode.HoverProvider {
     public provideHover(
         document: vscode.TextDocument,
         position: vscode.Position,
         token: vscode.CancellationToken
     ): vscode.Hover | null {
-        if (!Configuration.xqueryShowHovers()) return null;
+       
         const range = document.getWordRangeAtPosition(position);
         if (!range) return null
         const word = document.getText(range);
-        const linter =   diagnosticCollectionXQuery.xqlint(document.uri);     
-        const x=inspectAst(linter,position);
-        if (x.type==='WS') return null;
-        const parsePath = new vscode.MarkdownString(
-            '<a href="https://quodatum.github.io/basex-xqparse/i-BaseX.xhtml#'+x.path[0]+'">Path:<a>');
-        parsePath.appendText(x.path.join("/"));
-        parsePath.supportHtml = true;
-        parsePath.isTrusted = true;
+        const linter = xqLinters.xqlint(document.uri);
+        const what = inspectAst(linter, position);
+        if (what.type === 'WS') return null;
 
-        return new vscode.Hover([
-            `[${position.line},${position.character}] Type: ${x.name}, Word: ${word}`,
-            `value: ${x.value}, type: ${ x.type}`,
-            parsePath
-        ]);
+        const hovers:vscode.MarkdownString[]=[
+            what.display?markdownString("**"+what.display+"**"):null,
+            markdownString(what.get?.description)
+        ];
+
+        if (Configuration.xqueryShowHovers()) {
+            const parsePath = markdownString(
+                '<a href="https://quodatum.github.io/basex-xqparse/i-BaseX.xhtml#' + what.path[0] + '">Path:<a>');
+            parsePath.appendText(what.path.join("/"));
+            const debugInfo=`[${position.line},${position.character}] Word: ${word},
+            value: ${what.value}, type: ${what.type}`;
+            hovers.push(markdownString(debugInfo));
+            hovers.push( parsePath);
+        }
+        return new vscode.Hover(hovers);
     }
 }