+++ src/providers/symbols.ts
@@ -7,7 +7,7 @@ import {
 } from 'vscode';
 import { channel,importRange } from "../common";
 import { languageIds } from "../constants";
-import { diagnosticCollectionXQuery } from "../extension";
+import { xqLinters } from "../extension";
 //
 // This class handles XQuery Symbols
 //
@@ -47,7 +47,7 @@ export class DocumentSymbols implements DocumentSymbolProvider {
     channel.log("DocumentSymbols: " + document.uri);
     const symbols: DocumentSymbol[] = [];
 
-    const linter =   diagnosticCollectionXQuery.xqlint(document.uri); 
+    const linter =   xqLinters.xqlint(document.uri); 
 
     const xqdoc = linter.getXQDoc(true);
     channel.log("got xqdoc");