+++ src/xqlints.ts
@@ -11,19 +11,20 @@ import {
     channel, isNotXQDoc, unsupportedScheme,
     Configuration, importRange, findEditor
 } from "./common";
+import { TextDocument } from 'vscode';
 
 
 // DiagnosticCollection for XQuery documents
-export class XQLinter {
+export class XQLinters {
     diagnosticCollectionXQuery: vscode.DiagnosticCollection;
-    xqlintCollectionXQuery: Map<string, XQLint>;
+    xqlintCollection: Map<string, XQLint>;
 
     private diagEmitter = new vscode.EventEmitter<IXQParsedEvent>();
     onXQParsed: vscode.Event<IXQParsedEvent> = this.diagEmitter.event;
 
     constructor() {
         this.diagnosticCollectionXQuery = vscode.languages.createDiagnosticCollection(constants.diagnosticCollections.xquery);
-        this.xqlintCollectionXQuery = new Map();
+        this.xqlintCollection = new Map();
         channel.log("xqueryDiagnostics new");
     }
 
@@ -32,15 +33,19 @@ export class XQLinter {
         if (this.diagnosticCollectionXQuery.has(uri)) {
             channel.log("xqueryDiagnostics delete: " + uri);
             this.diagnosticCollectionXQuery.delete(uri);
-            this.xqlintCollectionXQuery.delete(uri.toString());
+            this.xqlintCollection.delete(uri.toString());
             this.diagEmitter.fire(new ParsedEvent(uri));
         }
     }
 
     clear(): void {
-        channel.log("xqueryDiagnostics clear");
-        this.diagnosticCollectionXQuery.clear();
-        this.xqlintCollectionXQuery.clear();
+        if (this) {
+            channel.log("xqueryDiagnostics clear");
+            this.xqlintCollection.clear();
+            this.diagnosticCollectionXQuery.clear();
+        } else {
+            channel.log("xqueryDiagnostics unset!");
+        }
     }
 
     get(uri: vscode.Uri): readonly vscode.Diagnostic[] {
@@ -51,63 +56,68 @@ export class XQLinter {
         this.diagnosticCollectionXQuery.set(uri, diagnostics);
     }
     has(uri: vscode.Uri): boolean {
-        return this.diagnosticCollectionXQuery.has(uri);
+        return this.xqlintCollection.has(uri.toString());
     }
 
-    update(uri: vscode.Uri, document: string) {
-        const xqlint = linter(uri, document);
-        this.xqlintCollectionXQuery.set(uri.toString(), xqlint);
-        this.diagEmitter.fire(new ParsedEvent(uri, xqlint))
+    update(doc: TextDocument) {
+        const xqlint = linter(doc.uri, doc.getText());
+        this.xqlintCollection.set(doc.uri.toString(), xqlint);
+        this.diagEmitter.fire(new ParsedEvent(doc.uri, xqlint))
     }
 
     xqlint(uri: vscode.Uri): XQLint {
-        const xq = this.xqlintCollectionXQuery.get(uri.toString());
+        const xq = this.xqlintCollection.get(uri.toString());
         if (xq) {
             return xq;
         } else {
-            throw { name: "NOXQLINT", message: "should not happen" };
+            throw { name: "NOXQLINT", message: uri.fsPath };
         }
     }
-
+    xqlint2(doc: TextDocument): XQLint {
+        if (!this.has(doc.uri)) {
+            this.update(doc);
+        }
+        return this.xqlintCollection.get(doc.uri.toString());
+    }
 }
+
+// do the parse
 function linter(uri: vscode.Uri, document: string) {
     const processor = Configuration.xqueryProcessor;
     const opts = { "processor": processor, "fileName": uri.fsPath };
     return new XQLint(document, opts);
+
 }
 
 /**
  * Analyzes the xquery document for problems. 
  * @param doc xquery document to analyze
- * @param xqueryDiagnostics diagnostic collection
+ * @param xqLinters diagnostic collection
  */
 export function refreshDiagnostics(doc: vscode.TextDocument,
-    xqueryDiagnostics: XQLinter,
+    xqLinters: XQLinters,
     reason: string): void {
     if (isNotXQDoc(doc)) return;
     const editor = findEditor(doc);
 
-    const isNew = !xqueryDiagnostics.has(doc.uri);
+    const isNew = !xqLinters.has(doc.uri);
     const refresh = reason === "change";
     channel.log((isNew ? "🆕" : "") + (refresh ? "♻️" : "") + "refreshDiagnostics " + reason + " " + doc.uri.toString());
     if (isNew || refresh) {
-        const text = doc.getText();
-        xqueryDiagnostics.update(doc.uri, text);
-        const xqlint = xqueryDiagnostics.xqlint(doc.uri);
-        const diagnostics = new Array<vscode.Diagnostic>();
        
+        xqLinters.update(doc);
+        const xqlint = xqLinters.xqlint(doc.uri);
+        const diagnostics = new Array<vscode.Diagnostic>();
+
         pushXQLintDiagnostics(diagnostics, xqlint);
 
-        xqueryDiagnostics.set(doc.uri, diagnostics);
+        xqLinters.set(doc.uri, diagnostics);
         channel.log("diagnostics.length: " + diagnostics.length);
     }
-    if (editor) {
-        const xqlint = xqueryDiagnostics.xqlint(doc.uri);
-        decorate(editor, xqlint);
-    }
+    
 }
 // forward doc changes
-export function subscribeToDocumentChanges(context: vscode.ExtensionContext, xqueryDiagnostics: XQLinter): void {
+export function subscribeToDocumentChanges(context: vscode.ExtensionContext, xqueryDiagnostics: XQLinters): void {
     const ed = vscode.window.activeTextEditor;
     if (ed && !unsupportedScheme(ed.document.uri)) {
         refreshDiagnostics(ed.document, xqueryDiagnostics, "current");
@@ -135,26 +145,20 @@ export function subscribeToDocumentChanges(context: vscode.ExtensionContext, xqu
 
 function pushXQLintDiagnostics(
     diagnostics: vscode.Diagnostic[],
-    linter: XQLint): vscode.Diagnostic[] {
+    linter: XQLint) {
+    linter.getMarkers().forEach(
+        (mark: Marker) => {
+            const type = mark.level === "error" ? vscode.DiagnosticSeverity.Error : vscode.DiagnosticSeverity.Warning
+            diagnostics.push(new vscode.Diagnostic(
+                importRange(mark.pos),
+                mark.message,
+                isSuppressed(mark.message) ? vscode.DiagnosticSeverity.Information : type
+            ));
+        }
+    );
+   
 
-    linter.getErrors().forEach((error: Marker) => {
-        diagnostics.push(new vscode.Diagnostic(
-            importRange(error.pos),
-            error.message,
-            isSuppressed(error.message) ? vscode.DiagnosticSeverity.Information : vscode.DiagnosticSeverity.Error
-        ));
-    });
-
-    linter.getWarnings().forEach((warning: Marker) => {
-        diagnostics.push(new vscode.Diagnostic(
-            importRange(warning.pos),
-            warning.message,
-            vscode.DiagnosticSeverity.Warning
-        ));
-    });
-    return diagnostics;
 }
-
 
 // [XQST0059] module "http://config" not found
 // [XPST0008] "list-details#0": undeclared function
@@ -162,35 +166,4 @@ function isSuppressed(msg: string): boolean {
     const errs = Configuration.xquerySuppressErrors;
     return errs.some((x) => msg.includes(x));
 }
-const parseFailedDecorationType = vscode.window.createTextEditorDecorationType({
-    borderWidth: '1px',
-    borderStyle: 'solid',
 
-    overviewRulerColor: 'red',
-    overviewRulerLane: vscode.OverviewRulerLane.Right,
-
-    light: {
-        // this color will be used in light color themes
-        borderColor: 'darkred',
-        backgroundColor: 'lightpink',
-    },
-    dark: {
-        // this color will be used in dark color themes
-        borderColor: 'lightred',
-        backgroundColor: 'deeppink',
-    }
-});
-
-
-function decorate(editor: vscode.TextEditor, xqlint: XQLint) {
-    const theDecorations: vscode.DecorationOptions[] = [];
-    if (xqlint.hasSyntaxError()) {
-        const r = xqlint.getErrors();
-        const range1 = importRange(r[0].pos);
-        const lastLine = editor.document.lineAt(editor.document.lineCount - 1);
-        const range = range1.with(range1.end, lastLine.range.end)
-        const decoration = { range: range, hoverMessage: 'Parse failed at line:' + range1.end.line };
-        theDecorations.push(decoration);
-    }
-    editor.setDecorations(parseFailedDecorationType, theDecorations);
-}